+++ crates/field/benches/byte_iteration.rs
@@ -3,8 +3,8 @@
 use std::iter::repeat_with;
 
 use binius_field::{
-	arch::ArchOptimal, byte_iteration::create_partial_sums_lookup_tables,
-	scalars_collection::PackedSlice, BinaryField128b, BinaryField1b, BinaryField8b, PackedField,
+	arch::ArchOptimal, byte_iteration::create_partial_sums_lookup_tables, packed::PackedSlice,
+	BinaryField128b, BinaryField1b, BinaryField8b, PackedField,
 };
 use criterion::{
 	criterion_group, criterion_main, measurement::WallTime, BenchmarkGroup, Criterion,
